+++ Library/Homebrew/os/mac/pathname.rb
@@ -1,3 +1,5 @@
+require "os/mac/shared_mach"
+
 class Pathname
   if ENV["HOMEBREW_RUBY_MACHO"]
     require "os/mac/ruby_mach"
@@ -6,4 +8,6 @@ class Pathname
     require "os/mac/cctools_mach"
     include CctoolsMachO
   end
+
+  include SharedMachO
 end
